+++ admin/reports/stock.class.php
@@ -10,6 +10,8 @@ if (!class_exists('WP_List_Table')) {
 
 class Jigoshop_Report_Stock extends WP_List_Table
 {
+	protected $max_items;
+
 	public function __construct()
 	{
 		parent::__construct(array(
@@ -43,8 +45,8 @@ class Jigoshop_Report_Stock extends WP_List_Table
 	{
 		$this->_column_headers = array($this->get_columns(), array(), $this->get_sortable_columns());
 		$current_page = absint($this->get_pagenum());
-		$per_page = apply_filters('jigoshop_admin_stock_report_products_per_page', 20);
-
+		$per_page = apply_filters('jigoshop_admin_stock_report_products_per_page', 1);
+		//$this->max_items
 		$this->get_items($current_page, $per_page);
 
 		$this->set_pagination_args(array(